refactor: User, Line 테스트코드 추가
